電話番号を保存できるように修正

+++ db/schema.rb
@@ -39,7 +39,7 @@ ActiveRecord::Schema.define(version: 2021_02_27_074131) do
     t.string "city", default: "", null: false
     t.string "house_number", default: "", null: false
     t.string "building_name", default: ""
-    t.integer "phone_number", null: false
+    t.bigint "phone_number", null: false
     t.bigint "purchase_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
